modified int with float

+++ esercizio1.py
@@ -1,4 +1,4 @@
 firstAddend = input("inserisci il primo addendo ");
 secondAddend = input("inserisci il secondo addendo ");
-total=int(firstAddend) * int(secondAddend)
+total=float(firstAddend) * float(secondAddend)
 print("il risultato è" + " " + str(total))
